backend/pyproject.toml
[project]
name = "dispatcher-agents"
version = "0.2.0"
description = "Voice Agent Conversation System - LiveKit + OpenAI Realtime"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "livekit>=0.12.0",
    "livekit-agents[openai]>=0.9.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

